feat: count on assets
